+++ RELEASENOTES.md
@@ -4,6 +4,7 @@
   - Enable trimming (reduces file size of release artifacts)
   - Allow interpolation to finish when game is paused and on level exit for smooth transitions
   - Fade sprites into max distance
+  - Autosave timer. Automatically write quicksave every x seconds.
 
 ## Bug fixes:
   - Fix per ammo values for box ammo and backpack amount from dehacked patch